Add files via upload
Added common functions file, refactored other files to use that. Have a feeling it will come in handy later, when we start to see what functions are getting reused a lot.

Finished out the logic for it to look for active games, and only call to update the g sheet when there is something to update. It'll exit the program if there is no schedule file available, or it is 4 hours after the last game start time. Also added a test mode in the configuration of that file to debug when we need to. It will take in a fake time, and let you run the script as if it is that time. It doesn't touch the schedule files or anything, but will call to update the google sheet based on fake time. I'm not sure whether that'll screw anything up if it isn't sunday, but we can deal w that later.