Add files via upload
Made the player data puller more resilient.
Added a player data cache, so overlapping players on teams will only be fetched once. This should be huge because we have so many overlapping teams. Hope this reduces the number of API limit/errors/timeouts we get.
Add a retry function, and error printing, to at least tell us when things have broken, instead of just exiting the program (at least I think this is what it is doing...)